user-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-dotenv
pyjwt
auth
requests